Use Next.JS and TypeScript
