+++ .prettierrc.json
@@ -1,0 +1,6 @@
+{
+    "tabWidth": 4,
+    "trailingComma": "all",
+    "singleQuote": false,
+    "jsxSingleQuote": false
+}
