+++ components/footer.tsx
@@ -1,0 +1,5 @@
+const Footer = () => {
+    return <div></div>;
+};
+
+export default Footer;
