+++ components/header.tsx
@@ -1,0 +1,5 @@
+const Header = () => {
+    return <div>Header!!!!!!</div>;
+};
+
+export default Header;
