+++ package.json
@@ -4,35 +4,39 @@
   "description": "<div id=\"top\"></div>",
   "private": true,
   "scripts": {
-    "start": "nodemon server",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "webpack -w"
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/NSC-Computer-Science-Club/club_website.git"
   },
   "author": "North Seattle Computer Science Club",
-  "license": "ISC",
+  "license": "MIT",
   "bugs": {
     "url": "https://github.com/NSC-Computer-Science-Club/club_website/issues"
   },
   "homepage": "https://github.com/NSC-Computer-Science-Club/club_website#readme",
   "dependencies": {
-    "jquery": "^3.6.0",
-    "nodemon": "^2.0.20",
-    "path": "^0.12.7",
-    "react": "^18.1.0",
-    "react-bootstrap": "^2.4.0",
-    "react-dom": "^18.1.0",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "^5.0.1"
+    "@chakra-ui/react": "^2.3.7",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@types/node": "^18.11.9",
+    "@typescript-eslint/eslint-plugin": "^5.42.1",
+    "@typescript-eslint/parser": "^5.42.1",
+    "eslint": "^8.27.0",
+    "eslint-config-next": "^13.0.2",
+    "eslint-plugin-react": "^7.31.10",
+    "framer-motion": "^7.6.5",
+    "next": "^13.0.2",
+    "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "typescript": "^4.8.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.18.2",
-    "@babel/preset-env": "^7.18.2",
-    "babel-loader": "^8.2.5",
-    "webpack": "^5.73.0",
-    "webpack-cli": "^4.9.2"
+    "@types/react": "18.0.25"
   }
 }
