+++ pages/index.tsx
@@ -1,0 +1,16 @@
+import Header from "../components/header";
+import Footer from "../components/footer";
+
+const Index = () => {
+    return (
+        <>
+            <Header />
+
+            {/* <Body /> */}
+
+            <Footer />
+        </>
+    );
+};
+
+export default Index;
