--- src/App.js
@@ -1,14 +1,0 @@
-import Header from './components/Header.js';
-import Footer from './components/Footer.js';
-
-function App() {
-    return(
-        <div>
-            <Header />
-            {/* <Body /> */}
-            <Footer />
-        </div>
-    )
-}
-
-export default App;
