--- src/components/footer.js
@@ -1,9 +1,0 @@
-import useBreakpoints from "./useBreakpoints.js";
-
-const Footer = () => {
-    return(
-        <div></div>
-    )
-}
-
-export default Footer;
