--- src/components/header.js
@@ -1,11 +1,0 @@
-import useBreakpoints from "./useBreakpoints.js";
-
-const Header = () => {
-    return(
-        <div>
-            Header!!!!!!
-        </div>
-    )
-}
-
-export default Header;
